@@ -1,11 +1,11 @@
 name: flutter_animated_dialog
 description: A new Flutter dialog with a series of beautiful animations, slide fade rotate size scale rotate3D animations.
-version: 1.2.1
+version: 2.0.2
 homepage: https://github.com/JackJonson/flutter_animated_dialog
 email: yswing1@gmail.com
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.18.1 <3.0.0"
 
 dependencies:
   flutter:
