@@ -67,17 +67,17 @@ enum DialogTransitionType {
 }
 
 /// Displays a Material dialog above the current contents of the app
-Future<T> showAnimatedDialog<T>({
-  @required BuildContext context,
+Future<T?> showAnimatedDialog<T extends Object?>({
+  required BuildContext context,
   bool barrierDismissible = false,
-  @required WidgetBuilder builder,
-  animationType = DialogTransitionType.fade,
+  required WidgetBuilder builder,
+  animationType = DialogTransitionType.size,
   Curve curve = Curves.linear,
-  Duration duration,
-  AlignmentGeometry alignment = Alignment.center,
-  Axis axis,
+  Duration? duration,
+  Alignment alignment = Alignment.center,
+  Color? barrierColor,
+  Axis? axis = Axis.horizontal,
 }) {
-  assert(builder != null);
   assert(debugCheckHasMaterialLocalizations(context));
 
   final ThemeData theme = Theme.of(context);
@@ -91,22 +91,19 @@ Future<T> showAnimatedDialog<T>({
       return SafeArea(
         // top: false,
         child: Builder(builder: (BuildContext context) {
-          return theme != null
-              ? Theme(data: theme, child: pageChild)
-              : pageChild;
+          return Theme(data: theme, child: pageChild);
         }),
       );
     },
     barrierDismissible: barrierDismissible,
     barrierLabel: MaterialLocalizations.of(context).modalBarrierDismissLabel,
-    barrierColor: Colors.black54,
+    barrierColor: barrierColor ?? Colors.black54,
     transitionDuration: duration ?? const Duration(milliseconds: 400),
     transitionBuilder: (BuildContext context, Animation<double> animation,
         Animation<double> secondaryAnimation, Widget child) {
       switch (animationType) {
         case DialogTransitionType.fade:
           return FadeTransition(opacity: animation, child: child);
-          break;
         case DialogTransitionType.slideFromRight:
           return SlideTransition(
             transformHitTests: false,
@@ -116,7 +113,6 @@ Future<T> showAnimatedDialog<T>({
             ).chain(CurveTween(curve: curve)).animate(animation),
             child: child,
           );
-          break;
         case DialogTransitionType.slideFromLeft:
           return SlideTransition(
             transformHitTests: false,
@@ -126,7 +122,6 @@ Future<T> showAnimatedDialog<T>({
             ).chain(CurveTween(curve: curve)).animate(animation),
             child: child,
           );
-          break;
         case DialogTransitionType.slideFromRightFade:
           return SlideTransition(
             position: Tween<Offset>(
@@ -138,7 +133,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.slideFromLeftFade:
           return SlideTransition(
             position: Tween<Offset>(
@@ -150,7 +144,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.slideFromTop:
           return SlideTransition(
             transformHitTests: false,
@@ -160,7 +153,6 @@ Future<T> showAnimatedDialog<T>({
             ).chain(CurveTween(curve: curve)).animate(animation),
             child: child,
           );
-          break;
         case DialogTransitionType.slideFromTopFade:
           return SlideTransition(
             position: Tween<Offset>(
@@ -172,7 +164,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.slideFromBottom:
           return SlideTransition(
             transformHitTests: false,
@@ -182,7 +173,6 @@ Future<T> showAnimatedDialog<T>({
             ).chain(CurveTween(curve: curve)).animate(animation),
             child: child,
           );
-          break;
         case DialogTransitionType.slideFromBottomFade:
           return SlideTransition(
             position: Tween<Offset>(
@@ -194,7 +184,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.scale:
           return ScaleTransition(
             alignment: alignment,
@@ -208,7 +197,6 @@ Future<T> showAnimatedDialog<T>({
             ),
             child: child,
           );
-          break;
         case DialogTransitionType.fadeScale:
           return ScaleTransition(
             alignment: alignment,
@@ -228,7 +216,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.scaleRotate:
           return ScaleTransition(
             alignment: alignment,
@@ -247,7 +234,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.rotate:
           return CustomRotationTransition(
             alignment: alignment,
@@ -255,7 +241,6 @@ Future<T> showAnimatedDialog<T>({
                 parent: animation, curve: Interval(0.0, 1.0, curve: curve))),
             child: child,
           );
-          break;
         case DialogTransitionType.fadeRotate:
           return CustomRotationTransition(
             alignment: alignment,
@@ -269,7 +254,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.rotate3D:
           return Rotation3DTransition(
             alignment: alignment,
@@ -285,10 +269,9 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.size:
           return Align(
-            alignment: alignment ?? Alignment.center,
+            alignment: alignment,
             child: SizeTransition(
               sizeFactor: CurvedAnimation(
                 parent: animation,
@@ -298,10 +281,9 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.sizeFade:
           return Align(
-            alignment: alignment ?? Alignment.center,
+            alignment: alignment,
             child: SizeTransition(
               sizeFactor: CurvedAnimation(
                 parent: animation,
@@ -316,10 +298,8 @@ Future<T> showAnimatedDialog<T>({
               ),
             ),
           );
-          break;
         case DialogTransitionType.none:
           return child;
-          break;
         default:
           return FadeTransition(opacity: animation, child: child);
       }
@@ -344,7 +324,7 @@ class CustomDialogWidget extends StatelessWidget {
   /// null, which implies a default that depends on the values of the other
   /// properties. See the documentation of [titlePadding] for details.
   const CustomDialogWidget({
-    Key key,
+    Key? key,
     this.title,
     this.titlePadding,
     this.titleTextStyle,
@@ -365,7 +345,7 @@ class CustomDialogWidget extends StatelessWidget {
   /// of the dialog.
   ///
   /// Typically a [Text] widget.
-  final Widget title;
+  final Widget? title;
 
   /// Padding around the title.
   ///
@@ -377,13 +357,13 @@ class CustomDialogWidget extends StatelessWidget {
   /// provided (but see [contentPadding]). If it _is_ null, then an extra 20
   /// pixels of bottom padding is added to separate the [title] from the
   /// [actions].
-  final EdgeInsetsGeometry titlePadding;
+  final EdgeInsetsGeometry? titlePadding;
 
   /// Style for the text in the [title] of this [AlertDialog].
   ///
   /// If null, [DialogTheme.titleTextStyle] is used, if that's null, defaults to
   /// [ThemeData.textTheme.title].
-  final TextStyle titleTextStyle;
+  final TextStyle? titleTextStyle;
 
   /// The (optional) content of the dialog is displayed in the center of the
   /// dialog in a lighter font.
@@ -392,7 +372,7 @@ class CustomDialogWidget extends StatelessWidget {
   /// message. As noted in the [AlertDialog] documentation, it's important
   /// to use a [SingleChildScrollView] if there's any risk that the content
   /// will not fit.
-  final Widget content;
+  final Widget? content;
 
   /// Padding around the content.
   ///
@@ -400,18 +380,18 @@ class CustomDialogWidget extends StatelessWidget {
   /// 20 pixels is provided above the content to separate the content from the
   /// title, and padding of 24 pixels is provided on the left, right, and bottom
   /// to separate the content from the other edges of the dialog.
-  final EdgeInsetsGeometry contentPadding;
+  final EdgeInsetsGeometry? contentPadding;
 
   /// Style for the text in the [content] of this [AlertDialog].
   ///
   /// If null, [DialogTheme.contentTextStyle] is used, if that's null, defaults
   /// to [ThemeData.textTheme.subhead].
-  final TextStyle contentTextStyle;
+  final TextStyle? contentTextStyle;
 
   /// The (optional) set of actions that are displayed at the bottom of the
   /// dialog.
   ///
-  /// Typically this is a list of [FlatButton] widgets.
+  /// Typically this is a list of [TextButton] widgets.
   ///
   /// These widgets will be wrapped in a [ButtonBar], which introduces 8 pixels
   /// of padding on each side.
@@ -419,17 +399,17 @@ class CustomDialogWidget extends StatelessWidget {
   /// If the [title] is not null but the [content] _is_ null, then an extra 20
   /// pixels of padding is added above the [ButtonBar] to separate the [title]
   /// from the [actions].
-  final List<Widget> actions;
+  final List<Widget>? actions;
 
   ///Widget in the bottom
-  final Widget bottomWidget;
+  final Widget? bottomWidget;
 
   /// {@macro flutter.material.dialog.backgroundColor}
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   /// {@macro flutter.material.dialog.elevation}
   /// {@macro flutter.material.material.elevation}
-  final double elevation;
+  final double? elevation;
 
   /// The semantic label of the dialog used by accessibility frameworks to
   /// announce screen transitions when the dialog is opened and closed.
@@ -442,13 +422,13 @@ class CustomDialogWidget extends StatelessWidget {
   ///
   ///  * [SemanticsConfiguration.isRouteName], for a description of how this
   ///    value is used.
-  final String semanticLabel;
+  final String? semanticLabel;
 
   /// {@macro flutter.material.dialog.shape}
-  final ShapeBorder shape;
+  final ShapeBorder? shape;
 
   ///Min width
-  final double minWidth;
+  final double? minWidth;
 
   @override
   Widget build(BuildContext context) {
@@ -456,16 +436,16 @@ class CustomDialogWidget extends StatelessWidget {
     final ThemeData theme = Theme.of(context);
     final DialogTheme dialogTheme = DialogTheme.of(context);
     final List<Widget> children = <Widget>[];
-    String label = semanticLabel;
+    String? label = semanticLabel;
 
     if (title != null) {
       children.add(Padding(
         padding: titlePadding ??
             EdgeInsets.fromLTRB(24.0, 24.0, 24.0, content == null ? 20.0 : 0.0),
         child: DefaultTextStyle(
-          style: titleTextStyle ??
+          style: (titleTextStyle ??
               dialogTheme.titleTextStyle ??
-              theme.textTheme.headline6,
+              theme.textTheme.headline6)!,
           child: Semantics(
             child: title,
             namesRoute: true,
@@ -481,18 +461,18 @@ class CustomDialogWidget extends StatelessWidget {
         case TargetPlatform.android:
         case TargetPlatform.fuchsia:
           label = semanticLabel ??
-              MaterialLocalizations.of(context)?.alertDialogLabel;
+              MaterialLocalizations.of(context).alertDialogLabel;
           break;
         case TargetPlatform.linux:
           label = semanticLabel ??
-              MaterialLocalizations.of(context)?.alertDialogLabel;
+              MaterialLocalizations.of(context).alertDialogLabel;
           break;
         case TargetPlatform.macOS:
           label = semanticLabel;
           break;
         case TargetPlatform.windows:
           label = semanticLabel ??
-              MaterialLocalizations.of(context)?.alertDialogLabel;
+              MaterialLocalizations.of(context).alertDialogLabel;
           break;
       }
     }
@@ -501,12 +481,12 @@ class CustomDialogWidget extends StatelessWidget {
       children.add(
         Flexible(
           child: Padding(
-            padding: contentPadding,
+            padding: contentPadding!,
             child: DefaultTextStyle(
-              style: contentTextStyle ??
+              style: (contentTextStyle ??
                   dialogTheme.contentTextStyle ??
-                  theme.textTheme.subtitle1,
-              child: content,
+                  theme.textTheme.subtitle1)!,
+              child: content!,
             ),
           ),
         ),
@@ -514,13 +494,13 @@ class CustomDialogWidget extends StatelessWidget {
     }
 
     if (bottomWidget != null) {
-      children.add(bottomWidget);
+      children.add(bottomWidget!);
     } else if (actions != null) {
       children.add(
         ButtonBarTheme(
           data: ButtonBarTheme.of(context),
           child: ButtonBar(
-            children: actions,
+            children: actions!,
           ),
         ),
       );
@@ -575,7 +555,7 @@ class CustomDialog extends StatelessWidget {
   ///
   /// Typically used in conjunction with [showDialog].
   const CustomDialog({
-    Key key,
+    Key? key,
     this.backgroundColor,
     this.elevation,
     this.insetAnimationDuration = const Duration(milliseconds: 100),
@@ -592,7 +572,7 @@ class CustomDialog extends StatelessWidget {
   ///
   /// If `null`, [ThemeData.cardColor] is used.
   /// {@endtemplate}
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   /// {@template flutter.material.dialog.elevation}
   /// The z-coordinate of this [Dialog].
@@ -601,22 +581,22 @@ class CustomDialog extends StatelessWidget {
   /// dialog's elevation is 24.0.
   /// {@endtemplate}
   /// {@macro flutter.material.material.elevation}
-  final double elevation;
+  final double? elevation;
 
   /// The duration of the animation to show when the system keyboard intrudes
   /// into the space that the dialog is placed in.
   ///
   /// Defaults to 100 milliseconds.
-  final Duration insetAnimationDuration;
+  final Duration? insetAnimationDuration;
 
   /// The curve to use for the animation shown when the system keyboard intrudes
   /// into the space that the dialog is placed in.
   ///
   /// Defaults to [Curves.fastOutSlowIn].
-  final Curve insetAnimationCurve;
+  final Curve? insetAnimationCurve;
 
   ///Min width of the dialog
-  final double minWidth;
+  final double? minWidth;
 
   /// {@template flutter.material.dialog.shape}
   /// The shape of this dialog's border.
@@ -625,12 +605,12 @@ class CustomDialog extends StatelessWidget {
   ///
   /// The default shape is a [RoundedRectangleBorder] with a radius of 2.0.
   /// {@endtemplate}
-  final ShapeBorder shape;
+  final ShapeBorder? shape;
 
   /// The widget below this widget in the tree.
   ///
   /// {@macro flutter.widgets.child}
-  final Widget child;
+  final Widget? child;
 
   // TODO(johnsonmh): Update default dialog border radius to 4.0 to match material spec.
   static const RoundedRectangleBorder _defaultDialogShape =
@@ -644,8 +624,8 @@ class CustomDialog extends StatelessWidget {
     return AnimatedPadding(
       padding: MediaQuery.of(context).viewInsets +
           const EdgeInsets.symmetric(horizontal: 40.0, vertical: 24.0),
-      duration: insetAnimationDuration,
-      curve: insetAnimationCurve,
+      duration: insetAnimationDuration!,
+      curve: insetAnimationCurve!,
       child: MediaQuery.removeViewInsets(
         removeLeft: true,
         removeTop: true,
