@@ -18,30 +18,29 @@ class Rotation3DTransition extends AnimatedWidget {
   ///
   /// The [turns] argument must not be null.
   const Rotation3DTransition({
-    Key key,
-    @required Animation<double> turns,
+    Key? key,
+    required Animation<double> turns,
     this.alignment = Alignment.center,
     this.child,
-  })  : assert(turns != null),
-        super(key: key, listenable: turns);
+  }) : super(key: key, listenable: turns);
 
   /// The animation that controls the rotation of the child.
   ///
   /// If the current value of the turns animation is v, the child will be
   /// rotated v * 2 * pi radians before being painted.
-  Animation<double> get turns => listenable;
+  Animation<double> get turns => listenable as Animation<double>;
 
   /// The alignment of the origin of the coordinate system around which the
   /// rotation occurs, relative to the size of the box.
   ///
   /// For example, to set the origin of the rotation to top right corner, use
   /// an alignment of (1.0, -1.0) or use [Alignment.topRight]
-  final Alignment alignment;
+  final Alignment? alignment;
 
   /// The widget below this widget in the tree.
   ///
   /// {@macro flutter.widgets.child}
-  final Widget child;
+  final Widget? child;
 
   @override
   Widget build(BuildContext context) {
@@ -73,30 +72,29 @@ class CustomRotationTransition extends AnimatedWidget {
   ///
   /// The [turns] argument must not be null.
   const CustomRotationTransition({
-    Key key,
-    @required Animation<double> turns,
+    Key? key,
+    required Animation<double> turns,
     this.alignment = Alignment.center,
     this.child,
-  })  : assert(turns != null),
-        super(key: key, listenable: turns);
+  }) : super(key: key, listenable: turns);
 
   /// The animation that controls the rotation of the child.
   ///
   /// If the current value of the turns animation is v, the child will be
   /// rotated v * 2 * pi radians before being painted.
-  Animation<double> get turns => listenable;
+  Animation<double> get turns => listenable as Animation<double>;
 
   /// The alignment of the origin of the coordinate system around which the
   /// rotation occurs, relative to the size of the box.
   ///
   /// For example, to set the origin of the rotation to top right corner, use
   /// an alignment of (1.0, -1.0) or use [Alignment.topRight]
-  final Alignment alignment;
+  final Alignment? alignment;
 
   /// The widget below this widget in the tree.
   ///
   /// {@macro flutter.widgets.child}
-  final Widget child;
+  final Widget? child;
 
   @override
   Widget build(BuildContext context) {
